Refactor mongo code in database package
